requirements.txt
flask
spacy==3.7.2
pdfminer.six
python-docx
pandas
gunicorn
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl